+++ example/src/demos/index.tsx
@@ -1,6 +1,6 @@
 import { lazy } from 'react'
 
-//const Viewcube = { descr: '', tags: [], Component: lazy(() => import('./Viewcube')), dev: true }
+const Viewcube = { descr: '', tags: [], Component: lazy(() => import('./Viewcube')), dev: true }
 const Selection = { descr: '', tags: [], Component: lazy(() => import('./Selection')), dev: true }
 const Pointcloud = { descr: '', tags: [], Component: lazy(() => import('./Pointcloud')), dev: true, bright: true }
 const Gestures = { descr: '', tags: [], Component: lazy(() => import('./Gestures')), bright: false }
@@ -28,6 +28,7 @@ const ResetProps = { descr: '', tags: [], Component: lazy(() => import('./ResetP
 //const Animation = { descr: '', tags: [], Component: lazy(() => import('./Animation')), dev: true, bright: true }
 const AutoDispose = { descr: '', tags: [], Component: lazy(() => import('./AutoDispose')), dev: true, bright: true }
 const Layers = { descr: '', tags: [], Component: lazy(() => import('./Layers')), dev: true, bright: true }
+const MultiMaterial = { descr: '', tags: [], Component: lazy(() => import('./MultiMaterial')), dev: true, bright: true }
 const Test = { descr: '', tags: [], Component: lazy(() => import('./Test')), dev: true, bright: true }
 const SuspenseAndErrors = {
   descr: '',
@@ -52,7 +53,9 @@ export {
   //Animation,
   AutoDispose,
   Layers,
-  Test,
+  MultiMaterial,
   SuspenseAndErrors,
   ContextMenuOverride,
+  Viewcube,
+  Test,
 }