+++ packages/fiber/src/core/hooks.ts
@@ -3,7 +3,6 @@ import * as React from 'react'
 import { StateSelector, EqualityChecker } from 'zustand'
 import type { GLTF } from 'three/examples/jsm/loaders/GLTFLoader'
 import { useAsset } from 'use-asset'
-
 import { context, RootState, RenderCallback } from './store'
 import { stringLiteral } from '@babel/types'
 
@@ -45,7 +44,7 @@ export function useFrame(callback: RenderCallback, renderPriority: number = 0): 
   const ref = React.useRef<RenderCallback>(callback)
   React.useLayoutEffect(() => void (ref.current = callback), [callback])
   // Subscribe on mount, unsubscribe on unmount
-  React.useLayoutEffect(() => subscribe(ref, renderPriority), [renderPriority])
+  React.useLayoutEffect(() => subscribe(ref, renderPriority), [renderPriority, subscribe])
   return null
 }