+++ packages/fiber/src/web/events.ts
@@ -20,10 +20,7 @@ export function createPointerEvents(store: UseStore<RootState>): EventManager<HT
 
   return {
     connected: false,
-    handlers: (Object.keys(names).reduce(
-      (acc, key) => ({ ...acc, [key]: handlePointer(key) }),
-      {},
-    ) as unknown) as Events,
+    handlers: Object.keys(names).reduce((acc, key) => ({ ...acc, [key]: handlePointer(key) }), {}) as unknown as Events,
     connect: (target: HTMLElement) => {
       const { set, events } = store.getState()
       events.disconnect?.()